refactor: smaller di files, assert text options, more au slot tests
