+++ packages/kernel/src/index.ts
@@ -3,14 +3,6 @@ export {
 } from './platform';
 
 export {
-  createResolver,
-  all,
-  factory,
-  type IAllResolver,
-  type IFactoryResolver,
-  type IOptionalResolver,
-  type IResolvedFactory,
-  type INewInstanceResolver,
   DI,
   IContainer,
   type IFactory,
@@ -20,11 +12,6 @@ export {
   type IResolver,
   IServiceLocator,
   type Key,
-  lazy,
-  type ILazyResolver,
-  type IResolvedLazy,
-  optional,
-  ignore,
   type RegisterSelf,
   type ResolveCallback,
   singleton,
@@ -35,10 +22,6 @@ export {
   InstanceProvider,
   type Resolved,
   type Transformer,
-  newInstanceForScope,
-  newInstanceOf,
-  ContainerConfiguration,
-  DefaultResolver,
   type IContainerConfiguration,
 } from './di';
 
@@ -46,11 +29,36 @@ export {
   resolve,
   type IResolvedInjection,
   Registrable,
+  ContainerConfiguration,
+  DefaultResolver,
 } from './di.container';
 
 export {
   Registration,
 } from './di.registration';
+
+export {
+  createResolver,
+  all,
+  factory,
+  type IAllResolver,
+  type IFactoryResolver,
+  type IOptionalResolver,
+  type IResolvedFactory,
+  type INewInstanceResolver,
+  lazy,
+  type ILazyResolver,
+  type IResolvedLazy,
+  optional,
+  ignore,
+  newInstanceForScope,
+  newInstanceOf,
+  type ICallableResolver,
+  allResources,
+  optionalResource,
+  own,
+  resource,
+} from './di.resolvers';
 
 export {
   type Class,
