+++ packages/kernel/src/utilities.ts
@@ -1,6 +1,7 @@
 import { Metadata } from '@aurelia/metadata';
 
 /** @internal */ export const objectFreeze = Object.freeze;
+/** @internal */ export const objectAssign = Object.assign;
 /** @internal */ export const safeString = String;
 /** @internal */ export const getOwnMetadata = Metadata.getOwn;
 /** @internal */ export const hasOwnMetadata = Metadata.hasOwn;
