+++ packages/runtime-html/src/utilities-di.ts
@@ -5,40 +5,9 @@ import {
   type Key,
   type Constructable,
   type IContainer,
-  type IAllResolver,
-  IOptionalResolver,
-  createResolver,
 } from '@aurelia/kernel';
 import { defineMetadata, getAnnotationKeyFor, getOwnMetadata } from './utilities-metadata';
 import { IResourceKind } from './resources/resources-shared';
-
-export const resource = <T extends Key>(key: T) =>
-  createResolver((key, handler, requestor) =>
-    requestor.has(key, false)
-      ? requestor.get(key)
-      : requestor.root.get(key))(key);
-
-export const optionalResource = <T extends Key>(key: T) =>
-  createResolver((key, handler, requestor) =>
-    (requestor.has(key, false)
-      ? requestor.get(key)
-      : requestor.root.has(key, false)
-        ? requestor.root.get(key)
-        : void 0))(key) as IOptionalResolver<T>;
-/**
- * A resolver builder for resolving all registrations of a key
- * with resource semantic (leaf + root + ignore middle layer container)
- */
-export const allResources = <T extends Key>(key: T) =>
-  createResolver((key, handler, requestor) => {
-    if (/* is root? */requestor.root === requestor) {
-      return requestor.getAll(key, false);
-    }
-
-    return requestor.has(key, false)
-      ? requestor.getAll(key, false).concat(requestor.root.getAll(key, false))
-      : requestor.root.getAll(key, false);
-  })(key) as IAllResolver<T>;
 
 /** @internal */
 export const createInterface = DI.createInterface;
